svm: move FileSystemProviderSupport to buildtimeinit
