+++ substratevm/src/com.oracle.svm.core/src/com/oracle/svm/core/jdk/runtimeinit/FileSystemProviderRuntimeInitSupport.java
@@ -40,6 +40,7 @@ import com.oracle.svm.core.annotate.TargetClass;
 import com.oracle.svm.core.feature.AutomaticallyRegisteredFeature;
 import com.oracle.svm.core.feature.InternalFeature;
 import com.oracle.svm.core.jdk.JDKInitializedAtRunTime;
+import com.oracle.svm.core.jdk.buildtimeinit.FileSystemProviderBuildTimeInitSupport;
 import com.oracle.svm.core.jdk.resources.NativeImageResourceFileSystemProvider;
 import com.oracle.svm.core.util.BasedOnJDKFile;
 import com.oracle.svm.core.util.VMError;
@@ -47,13 +48,13 @@ import com.oracle.svm.util.ReflectionUtil;
 
 /**
  * This file contains substitutions that are required for initializing {@link FileSystemProvider} at
- * image run time. Other related functionality (general and build time initialization) can be found
- * in {@link com.oracle.svm.core.jdk.FileSystemProviderSupport}.
+ * image {@linkplain JDKInitializedAtRunTime run time}. Other related functionality (general and
+ * build time initialization) can be found in {@link FileSystemProviderBuildTimeInitSupport}.
  *
  * @see JDKInitializedAtRunTime
- * @see com.oracle.svm.core.jdk.FileSystemProviderSupport
+ * @see FileSystemProviderBuildTimeInitSupport
  */
-final class FileSystemProviderRuntimeInitSupport {
+public final class FileSystemProviderRuntimeInitSupport {
 }
 
 @AutomaticallyRegisteredFeature
